Move UTXO can_refund_htlc to utxo_common and impl it for QTUM as well.
